# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HashWatch API",
    "description": "Backend API for HashWatch \u2014 monitor and manage cryptocurrency miners in real-time.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://{host}:{port}/api",
      "variables": {
        "host": {
          "default": "localhost",
          "description": "API host"
        },
        "port": {
          "default": "4000",
          "description": "API port"
        }
      }
    }
  ],
  "paths": {
...
    "/miners": {
      "get": {
        "tags": ["Miners"],
        "summary": "List all miners for the authenticated user",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of miners",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Miner"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Miners"],
        "summary": "Register a new miner",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMinerRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Miner created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Miner"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/miners/{id}": {
      "put": {
        "tags": ["Miners"],
        "summary": "Update a miner (partial)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMinerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Miner updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Miner"
                }
              }
            }
          },
          "400": {
            "description": "No fields to update",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Miner not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Miners"],
        "summary": "Delete a miner",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Miner deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "404": {
            "description": "Miner not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/miners/pools": {
      "get": {
        "tags": ["Miners"],
        "summary": "List pool statistics for all miners",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Array of pool stats",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PoolStat"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/miners/{minerId}/notes": {
      "get": {
        "tags": ["Miner Notes"],
        "summary": "List notes for a miner",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "minerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of notes (newest first)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MinerNote"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Miner Notes"],
        "summary": "Add a note to a miner",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "minerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNoteRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Note created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MinerNote"
                }
              }
            }
          },
          "400": {
            "description": "Text is required (1-500 chars)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Miner not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/miners/{minerId}/notes/{noteId}": {
      "put": {
        "tags": ["Miner Notes"],
        "summary": "Update a note's text",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "minerId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "noteId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNoteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Note updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MinerNoteItem" }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Note not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Miner Notes"],
        "summary": "Delete a note from a miner",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "minerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "noteId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Note deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "404": {
            "description": "Note not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /auth/register or /auth/login. Expires in 30 days."
      }
    },
    "schemas": {
...
      "Miner": {
        "type": "object",
        "required": ["id", "userId", "name", "ip", "port", "addedat"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userid": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "ip": {
            "type": "string",
            "format": "ip"
          },
          "port": {
            "type": "integer",
            "default": 80
          },
          "addedat": {
            "type": "string",
            "format": "date-time"
          },
          "lastseen": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "CreateMinerRequest": {
        "type": "object",
        "required": ["name", "ip"],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50
          },
          "ip": {
            "type": "string",
            "format": "ip"
          },
          "port": {
            "type": "integer",
            "default": 80
          }
        }
      },
      "UpdateMinerRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50
          },
          "ip": {
            "type": "string",
            "format": "ip"
          },
          "port": {
            "type": "integer"
          }
        },
        "minProperties": 1
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "DeleteResponse": {
        "type": "object",
        "properties": {
          "deleted": {
            "type": "boolean",
            "enum": [true]
          }
        }
      },
...
      "PoolStat": {
        "type": "object",
        "properties": {
          "pool": {
            "type": "string"
          },
          "minerCount": {
            "type": "integer"
          },
          "totalHashrate": {
            "type": "number"
          },
          "totalShares": {
            "type": "integer"
          },
          "bestDiff": {
            "type": "string"
          },
          "acceptRate": {
            "type": "number"
          }
        }
      },
      "MinerNote": {
        "type": "object",
        "required": ["id", "minerId", "userId", "text", "createdAt"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "minerId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "text": {
            "type": "string"
          },
          "createdAt": {
            "type": "integer",
            "description": "Unix ms timestamp"
          }
        }
      },
      "CreateNoteRequest": {
        "type": "object",
        "required": ["text"],
        "properties": {
          "text": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          }
        }
      },
...
    }
  },
...
}